core/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
redis==5.0.4
PyJWT==2.9.0
starlette==0.37.2
httpx==0.27.0
python-multipart==0.0.9
